Move recent metamanager changes to ts
